+++ lib/gn_crossmap.rb
@@ -67,11 +67,13 @@ module GnCrossmap
     private
 
     def create_resolver(writer, opts)
-      Resolver.new(writer, opts.data_source_id, opts.resolver_url, opts.stats)
+      Resolver.new(writer, opts.data_source_id, opts.resolver_url,
+                   opts.stats, opts.with_classification)
     end
 
     def create_writer(reader, output_io, opts)
-      Writer.new(output_io, reader.original_fields, output_name(opts.output))
+      Writer.new(output_io, reader.original_fields,
+                 output_name(opts.output), opts.with_classification)
     end
 
     def create_reader(input_io, opts)