+++ lib/gn_crossmap/writer.rb
@@ -3,9 +3,11 @@
 module GnCrossmap
   # Saves output from GN Resolver to disk
   class Writer
-    def initialize(output_io, original_fields, output_name)
+    def initialize(output_io, original_fields, output_name,
+                   with_classification = false)
       @output_io = output_io
       @output_fields = output_fields(original_fields)
+      @output_fields << :classification if with_classification
       @output = CSV.new(@output_io, col_sep: "\t")
       @output << @output_fields
       @output_name = output_name
@@ -24,7 +26,7 @@ module GnCrossmap
     private
 
     def output_fields(original_fields)
-      original_fields + %i[matchedType inputName matchedName
+      original_fields + %i[matchedType inputName matchedName inputCanonicalForm
                            matchedCanonicalForm inputRank matchedRank
                            synonymStatus acceptedName matchedEditDistance
                            matchedScore matchTaxonID]